+++ src/main/java/com/umulam/fleen/health/repository/jpa/BusinessJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface BusinessJpaRepository extends JpaRepository<Business, Integer> {
+public interface BusinessJpaRepository extends JpaRepository<Business, Long> {
 
   @Query("SELECT b FROM Business b WHERE b.member.emailAddress = :emailAddress")
   Optional<Business> findBusinessByEmailAddress(String emailAddress);
